Add files via upload
Overview
In the exercise Creating an Executive Data Summary, you were asked to put into practice a range of formatting and calculation techniques that you have learned over the past couple of weeks. 

Your task in this exercise was to create an executive summary of a business’s month-by-month profit margin performance for Quarter 1 compared to the same period in the previous year. 

You were required to make formatting changes that would make the worksheet easier to read and ensure the requested analysis results stand out. In addition, you were asked to create formulas that had to: 

Produce customized quarter-one monthly totals across two business years using the SUMIF function.

Calculate the percentage difference for each month in quarter 1, 2023, from the same period in 2022.

And use a logical function to test the order value and display the correct tax amount.

This reading provides you with a step-by-step guide for identifying these results.